requirements.txt
spotipy==2.23.0
google-api-python-client==2.95.0
google-auth-oauthlib==1.0.0
google-auth==2.23.4
python-dotenv==1.0.1